+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v1.0 includes:
 * Updated README to detail how to use action.
 * Updated download dependencies to download libraries into separate folders separated by using UUIDs.
 * Soft crash on downloading dependencies so that duplicates don't crash the entire action.
+* Checking for detached head (pull requests) for git commands, so that the action doesn't crash.
 
 v0.14 includes:
 * Added default template in packages.